refactor: moved test layer creation/deletion to layer_harness
Moved from precomputed so that testing code can live in its own
space.

Also changed layer_delete to not ignore errors.

Added function create_storage in anticipation of additional tests being added.
