+++ python/test/test_precomputed.py
@@ -6,30 +6,8 @@ import numpy as np
 from neuroglancer.pipeline import Storage, Precomputed
 from neuroglancer.pipeline.task_creation import (upload_build_chunks, create_info_file_from_build,
     create_ingest_task, MockTaskQueue)
+from test.layer_harness import delete_layer, create_layer
 
-def create_layer(size, offset=[0,0,0], layer_type="image", layer_name="layer"):
-    storage = Storage('file:///tmp/removeme/'+ layer_name, n_threads=0)
-
-    if layer_type == "image":
-        random_data = np.random.randint(255, size=size, dtype=np.uint8)
-        upload_build_chunks(storage, random_data, offset)
-        # Jpeg encoding is lossy so it won't work
-        create_info_file_from_build(storage, layer_type= 'image', encoding="raw", force_chunk=[64,64,64])
-    elif layer_type == "affinities":
-        random_data = np.random.uniform(size=size).astype(np.float32)
-        upload_build_chunks(storage, random_data, offset)
-        create_info_file_from_build(storage, layer_type= 'image', encoding="raw", force_chunk=[64,64,64])
-    elif layer_type == "segmentation":
-        random_data = np.random.randint(0xFFFFFF, size=size, dtype=np.uint32)
-        upload_build_chunks(storage, random_data, offset)
-        # Jpeg encoding is lossy so it won't work
-        create_info_file_from_build(storage, layer_type= 'segmentation', encoding="raw", force_chunk=[64,64,64])
-    create_ingest_task(storage, MockTaskQueue())
-    return storage, random_data
-
-def delete_layer(layer_name="layer"):
-    shutil.rmtree("/tmp/removeme/"+layer_name, ignore_errors=True)
-    
 def test_aligned_read():
     delete_layer()
     storage, data = create_layer(size=(50,50,50,1), offset=(0,0,0))
